IDB: Implement IDBCursor.delete() <rdar://problem/15944203> and https://bugs.webkit.org/show_bug.cgi?id=127882
Reviewed by Sam Weinig.

IDBCursor.delete() actually relies on deleteRange(), which was implemented earlier.
So the only thing keeping it from working was a "Cursors iterate past their end" bug,
causing the delete transaction to be aborted.

* DatabaseProcess/IndexedDB/sqlite/SQLiteIDBCursor.cpp:
(WebKit::SQLiteIDBCursor::advanceOnce): When the cursor does the final iteration,
  clear the previous values to indicate completion.

* DatabaseProcess/IndexedDB/sqlite/SQLiteIDBTransaction.cpp:
(WebKit::SQLiteIDBTransaction::clearCursors): Noticed this opportunistic cleanup to
  iterate over unique_ptr<>s with an auto& instead of an auto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163252 268f45cc-cd09-0410-ab3c-d52691b4dbfc