WebKit Bot Watcher's Dashboard: Teach JSON.load() to interpret third argument as either an option dictionary or a failure callback https://bugs.webkit.org/show_bug.cgi?id=128080
Reviewed by Alexey Proskuryakov.

Currently JSON.load() takes a failure callback function as its third argument and
an option dictionary as its fourth argument. So, a caller that wants to ignore errors
and pass an option dictionary must pass null or a reference to an empty function for
the value of third argument. Instead, we should make the third argument polymorphic for
convenience. Then a caller can either pass a failure callback function or an option dictionary.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/Utilities.js:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163313 268f45cc-cd09-0410-ab3c-d52691b4dbfc