Remove CLASS_IF_GCC workarounds https://bugs.webkit.org/show_bug.cgi?id=128207
Reviewed by Anders Carlsson.

Remove the CLASS_IF_GCC macro that was defined to 'class' when using the GCC compiler.
The macro was then used in class friendship declarations for templated classes to avoid
corner-case compiler failures on both GCC pre-4.7 and MSVC pre-2013. The problematic
versions of both compilers are no longer supported, so this macro is good to go.

Source/JavaScriptCore: 

* heap/HeapBlock.h:
* heap/Region.h:

Source/WebCore: 

* bindings/generic/RuntimeEnabledFeatures.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163436 268f45cc-cd09-0410-ab3c-d52691b4dbfc