Respect SVG fragment identifiers in <img> src attribute https://bugs.webkit.org/show_bug.cgi?id=129387
Reviewed by Antti Koivisto.

Source/WebCore:

Test: svg/css/svg-resource-fragment-identifier-img-src.html

When providing an SVG image for a given renderer, check that the URL used to load
that image is taken into account in case it featured a fragment identifier, ensuring
that the CSS :target pseudo-class is correctly handled for SVG resources. This patch
is specific to <img> elements, specific support will also need to be added for various
CSS properties that support SVG images.

* svg/graphics/SVGImageCache.cpp:
(WebCore::SVGImageCache::imageForRenderer):
Check if the provided renderer is attached to an <img> element and, if so, pass the
resolved <img> source URL, taking into account srcset, to the SVGImageForContainer.

* svg/graphics/SVGImageForContainer.cpp:
(WebCore::SVGImageForContainer::setURL):
Trigger the FrameView machinery to ensure that the :target pseudo-class is respected
should the provided URL feature a fragment identifier.

* svg/graphics/SVGImageForContainer.h:
Declare the new setURL() method.

LayoutTests:

Test that we correctly handle the fragment identifier used in SVG URLs in <img> elements,
checking for correct srcset handling as well.

* svg/css/resources/fragment-identifiers.svg: Added.
* svg/css/svg-resource-fragment-identifier-img-src-expected.html: Added.
* svg/css/svg-resource-fragment-identifier-img-src.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164804 268f45cc-cd09-0410-ab3c-d52691b4dbfc