[iOS] DumpRenderTree Perl Support may build against wrong SDK and toolchain https://bugs.webkit.org/show_bug.cgi?id=129163
Reviewed by David Kilzer.

Fixes an issue where DumpRenderTree Perl Support may build with the specified
iOS Simulator SDK and toolchain instead of using the OS X SDK and toolchain.
This issue manifests itself in a build failure when building the IPhoneSimulatorNotification
bundle.

Currently we build DumpRenderTree Perl Support with respect to the variable SDKROOT,
which is specified either explicitly when building with MAKE(1) or implicitly by
build-{dumprendertree, webkit}. Instead we should always build Perl Support using
the OS X SDK and toolchain since the built products are only applicable to OS X.

* DumpRenderTree/ios/PerlSupport/Makefile:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164865 268f45cc-cd09-0410-ab3c-d52691b4dbfc