Some small modernizing tweaks to the coding style guide https://bugs.webkit.org/show_bug.cgi?id=129427
Reviewed by Alexey Proskuryakov.

My intent here was originally to do the nullptr style guide update, but that was already done.
I did some other improvements. There's still quite a bit to do -- the guide is out of date in
a number of subtle ways.

* coding/coding-style.html: Change "left side operator" example to be a more modern example
without non-WebKit-style abbreviations and such. Use auto& instead of const auto& in for loop
example, since that's normally preferred. Changed section title to say "zero" instead of "0".
Use words instead of abbreviations in code examples. Show that a modern C++ for loop is preferred
over index iteration in vector iteration example and also showed use of unsigned rather than
size_t since that is almost always what we want. Eliminated the use of PassOwnPtr in the "return
a newly created object" example. Use references instead of pointers in one example.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164930 268f45cc-cd09-0410-ab3c-d52691b4dbfc