Comment in RenderStyle is no longer accurate https://bugs.webkit.org/show_bug.cgi?id=129956
Reviewed by Simon Fraser.

RenderStyle::createStyleInheritingFromPseudoStyle was moved from RenderImage. Before it
was moved, there was an image-specific comment in the function. However, that comment
is both irrelevant to RenderStyle, and incorrect as the function is necessary for
RenderQuotes as well.

No new tests are necessary because there is no behavior change.

* rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::createStyleInheritingFromPseudoStyle):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165324 268f45cc-cd09-0410-ab3c-d52691b4dbfc