Only export initializeAndGetJSCLLVMAPI from libllvmForJSC.dylib https://bugs.webkit.org/show_bug.cgi?id=130224
Patch by Juergen Ributzka <juergen@apple.com> on 2014-03-13
Reviewed by Filip Pizlo.

This limits the exported symbols to only initializeAndGetJSCLLVMAPI from
the LLVM dylib. This allows the dylib to be safely used with other LLVM
dylibs on the same system. It also reduces the dynamic linking overhead
and also reduces the size by 6MB, because the linker can now dead strip
many unused functions.

* Configurations/LLVMForJSC.xcconfig:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165600 268f45cc-cd09-0410-ab3c-d52691b4dbfc