Incorrect behavior when mutating a typed array during set. https://bugs.webkit.org/show_bug.cgi?id=130428
Reviewed by Geoffrey Garen.

Source/JavaScriptCore:

This fixes a null derefence that occurs if a typed array
is mutated during the set() operation. The patch gets rid
of the "Quickly" version of setIndex that is assigning
JSValues of unknown type, as the numeric conversion can trigger
side effects that lead to neutering, and so we deref null.

* runtime/JSGenericTypedArrayView.h:
(JSC::JSGenericTypedArrayView::setIndex):
* runtime/JSGenericTypedArrayViewInlines.h:
(JSC::JSGenericTypedArrayView<Adaptor>::set):
(JSC::JSGenericTypedArrayView<Adaptor>::putByIndex):

LayoutTests:

Test behaviour of TypedArray.set(array) when we neuter the array.

* js/typed-array-mutated-during-set-expected.txt: Added.
* js/typed-array-mutated-during-set.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165989 268f45cc-cd09-0410-ab3c-d52691b4dbfc